+++ changelog.md
@@ -1,3 +1,4 @@
 ### General
 
 - ***Added*** carpet rule `tickFluids`
+- Added command `/ad xaero <id>` to share address as Xaero waypoint